Revert "Move dynamic title change of tab to BaseSessionWizardView"
This reverts commit a863fbd72803daab5a2a9db9f11a192890b42f1e.
